Extract a 1D water column from NEMO-MEDUSA

These functions read in target variables from NEMO-MEDUSA model outputs and return a water column.

Usage

get_grid_T_1D(filename, date, start, count)

get_ptrc_T_1D(filename, date, start, count)

get_grid_W_1D(filename, date, start, count)

get_grid_V_1D(filename, date, start, count)

get_grid_U_1D(filename, date, start, count)

Arguments

filename

The name of a netcdf file containing the title variables.

date

The numeric string representing the date contained by the data file.

start

an optional vector of indices to start subsetting at. See ncdf4 documentation.

count

an optional vector of steps to subset along. See ncdf4 documentation.

...

soaks up unused function arguments passed by the wrapper functions handling file architecture.

Value

A dataframe containing estimates of all variables at all depths for a single point in the model grid

Details

Each variable of interest in the netcdf file is imported, only reading at a single specified pixel for all depths.

Each model output file type contains a different set of variables, extracted by the relevant function variant:

File type 	Variables

grid_T_	Salinity, temperature, sea ice concentration.

grid_U_	Zonal currents.

grid_V_	Meridional currents.

grid_W_	Vertical velocity, vertical eddy diffusivitiy.

ptrc_T_	DIN, phytoplankton nitrogen content, phytoplankton chlorophyll, <br> zooplankton nitrogen content, detritus, silicate.

Some function variants have different arguments:

grid_W_ expects it's own scheme_w as depth levels are different between these and other files.

See also