Condense Daily netcdf Files into a Monthly Summary

This function takes the metadata for multiple netcdf files. It then creates a common spatial summary for all the variables of interest for a month.

whole_month_old(data, crop, grid, ...)

Arguments

data
A dataframe containing the metadata of multiple netcdf files from a common month.

...
Additional arguments to be passed to get_* functions.

Value

The function returns a dataframe containing the monthly average shalllow and deep spatial grids for all the variables of interest in NEMO-MEDUSA outputs.

Details

The function takes the metadata for a collection of files which contain data from the same month. The files are split into data packets which share the same file type, before being passed to `type_in_month` to be summarised. `type_in_month` reduces the NEMO-MEDUSA model outputs from large arrays to effectively two matrices. The summaries for each file type are returned to this function and get bound into a single dataframe. Points outside the project window are removed before saving the dataframe for the month in "./Objects/Months/".

Creating intermediate monthly objects allows the terabytes of data to be reduced without encountering memory issues. Also, working on independent monthly packets of data means we can parallelise any data processing for speed.

See also